Dump environment variables in alphabetical order.
Sorting makes output deterministic and therefore makes the "option/r"
test always work.

Indeed, to avoid hash attacks, there is extra randomization done in Perl
that makes the order of the keys in the hash tables non predictable.

If we do not force sorting in the output, we can have a different order
among the variables dumped, and that was making the "option/r" test
randomly fail since it is comparing output from two different invocations.