Decode all headers as UTF-8 for selection.
This allows transparent filtering of encoded headers with a
simple UTF-8 regular expression, without having to know about the
original header encoding (which might not have been UTF-8).

Naturally, this assumes that mailagent rules are held in a file which is
encoded in UTF-8 (should be the norm in 2023!) or whose regular
expression explicitly attempt to match UTF-8 characters.

If not, one can use \N{} escapes to specify Unicode character names or
codepoints.