Fit queued file names to our hardwired max length.
On modern systems, the PID value is no longer a 16-bit value hence we
can have very long file names on long-running systems.

To make things fit the mailagent queue listing format we decided to use
a long time ago, artificially cap the "PID" to being a number with at
most 5 digits.

Otherwise, queue listing becomes unreliable due to filename truncation
and the options/l test can fail because it may not be able to match
the actual files being queued with the output of "mailagent -l".